requirements.txt
aiogram==3.16.0
pandas==2.2.3
openpyxl==3.1.5
numpy==2.1.3
scikit-learn==1.5.2
SQLAlchemy==2.0.36
aiosqlite==0.20.0
python-dotenv==1.0.1
aiofiles==24.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
greenlet==3.1.1

